coverage: fixed missing "config-3.py" issue
This file is a configuration file which is shipped with certain versions
of the "opencv-python" package. Without this workaround ignoring
both "./config.py " and "./config-3.py", coverage is trying to find
those files in the current directory and it fails with the following
error message:
No source for code: 'C:\Dev\Projets\DataLab_Plugin_DPLMJ\config-3.py'.