Removed references to mypy (>pylance)
pylance already integrates a subset of pyright features
The idea is to switch pylance typing analysis to "basic" from time
to time, in order to show warning messages.
Most of the time however, it is quite drastic to fix all warnings,
so it should stay disabled on a daily basis.